Left Hand Shoulder patch via Fika
